agilityplugin: Rename plugin package to agility
