+++ runelite-client/src/main/java/net/runelite/client/plugins/agility/AgilityConfig.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.agilityplugin;
+package net.runelite.client.plugins.agility;
 
 import java.awt.Color;
 import net.runelite.client.config.Config;
