+++ runelite-client/src/main/java/net/runelite/client/plugins/agility/AgilityOverlay.java
@@ -23,7 +23,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.agilityplugin;
+package net.runelite.client.plugins.agility;
 
 import java.awt.Color;
 import static java.awt.Color.RED;
