+++ runelite-client/src/main/java/net/runelite/client/plugins/agility/Obstacles.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.agilityplugin;
+package net.runelite.client.plugins.agility;
 
 import com.google.common.collect.Sets;
 import java.util.Set;
